requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
aiohttp==3.9.3
asgiref==3.7.2
uvicorn==0.27.0 